feat: new opt config.OnlyColorPrefix
BREAKING CHANGE: this commit introduces the OnlyColorPrefix option, which allows the user to specify wheter or not to color the whole line of stdout+stderr of the process in its color. If it is set to true only the prefix (xx |) is highlighted in the processes color, if it is set to false the whole line is colored. By default this option is enabled to keep the output simple and clean, therefore this commit introduces a breaking change and requires a major version increment in the next release